+++ ft_numeric.c
@@ -43,14 +43,14 @@ void	ft_di(va_list *list_args, t_flags *flags)
 
 void	ft_u(va_list *list_args, t_flags *flags)
 {
-	unsigned long int	nb;
+	unsigned int	nb;
 	int					len;
 	int					rest;
 
 	rest = 0;
 	ft_flags(list_args, flags);
 	flags->opt = 2;
-	nb = va_arg(*list_args, unsigned long int);
+	nb = va_arg(*list_args, unsigned int);
 	len = ft_nblenx(nb, 10);
 	if ((rest = ft_flagsaffin(flags, len, nb, 0)) < 0)
 		return ;
